use output msbuild task parameters
