+++ src/coverlet.msbuild.tasks/CoverageResultTask.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Text;
+
 using ConsoleTables;
 using Coverlet.Core;
 using Coverlet.Core.Enums;
@@ -13,47 +14,25 @@ namespace Coverlet.MSbuild.Tasks
 {
     public class CoverageResultTask : Task
     {
-        private string _output;
-        private string _format;
-        private double _threshold;
-        private string _thresholdType;
-        private string _thresholdStat;
         private MSBuildLogger _logger;
 
         [Required]
-        public string Output
-        {
-            get { return _output; }
-            set { _output = value; }
-        }
+        public ITaskItem InstrumenterState { get; set; }
 
         [Required]
-        public string OutputFormat
-        {
-            get { return _format; }
-            set { _format = value; }
-        }
+        public string Output { get; set; }
 
         [Required]
-        public double Threshold
-        {
-            get { return _threshold; }
-            set { _threshold = value; }
-        }
+        public string OutputFormat { get; set; }
 
         [Required]
-        public string ThresholdType
-        {
-            get { return _thresholdType; }
-            set { _thresholdType = value; }
-        }
+        public double Threshold { get; set; }
 
         [Required]
-        public string ThresholdStat
-        {
-            get { return _thresholdStat; }
-            set { _thresholdStat = value; }
-        }
+        public string ThresholdType { get; set; }
+
+        [Required]
+        public string ThresholdStat { get; set; }
 
         public CoverageResultTask()
         {
@@ -66,10 +45,21 @@ namespace Coverlet.MSbuild.Tasks
             {
                 Console.WriteLine("\nCalculating coverage result...");
 
-                var coverage = InstrumentationTask.Coverage;
-                var result = coverage.GetCoverageResult();
+                if (InstrumenterState is null || !File.Exists(InstrumenterState.ItemSpec))
+                {
+                    _logger.LogError("Instrumenter result file not found");
+                    return false;
+                }
 
-                var directory = Path.GetDirectoryName(_output);
+                // Deserialize and cleanup
+                IInstrumentStateSerializer serializer = new JsonInstrumentStateSerializer();
+                InstrumenterState instrumenterState = serializer.Deserialize(new FileStream(InstrumenterState.ItemSpec, FileMode.Open));
+                File.Delete(InstrumenterState.ItemSpec);
+
+                ICoverageCalculator coverageResult = new CoverageCalculator(instrumenterState, _logger);
+                CoverageResult result = coverageResult.GetCoverageResult();
+
+                var directory = Path.GetDirectoryName(Output);
                 if (directory == string.Empty)
                 {
                     directory = Directory.GetCurrentDirectory();
@@ -79,7 +69,7 @@ namespace Coverlet.MSbuild.Tasks
                     Directory.CreateDirectory(directory);
                 }
 
-                var formats = _format.Split(',');
+                var formats = OutputFormat.Split(',');
                 foreach (var format in formats)
                 {
                     var reporter = new ReporterFactory(format).CreateReporter();
@@ -97,7 +87,7 @@ namespace Coverlet.MSbuild.Tasks
                     else
                     {
                         // Output to file
-                        var filename = Path.GetFileName(_output);
+                        var filename = Path.GetFileName(Output);
                         filename = (filename == string.Empty) ? $"coverage.{reporter.Extension}" : filename;
                         filename = Path.HasExtension(filename) ? filename : $"{filename}.{reporter.Extension}";
 
@@ -110,7 +100,7 @@ namespace Coverlet.MSbuild.Tasks
                 var thresholdTypeFlags = ThresholdTypeFlags.None;
                 var thresholdStat = ThresholdStatistic.Minimum;
 
-                foreach (var thresholdType in _thresholdType.Split(',').Select(t => t.Trim()))
+                foreach (var thresholdType in ThresholdType.Split(',').Select(t => t.Trim()))
                 {
                     if (thresholdType.Equals("line", StringComparison.OrdinalIgnoreCase))
                     {
@@ -126,11 +116,11 @@ namespace Coverlet.MSbuild.Tasks
                     }
                 }
 
-                if (_thresholdStat.Equals("average", StringComparison.OrdinalIgnoreCase))
+                if (ThresholdStat.Equals("average", StringComparison.OrdinalIgnoreCase))
                 {
                     thresholdStat = ThresholdStatistic.Average;
                 }
-                else if (_thresholdStat.Equals("total", StringComparison.OrdinalIgnoreCase))
+                else if (ThresholdStat.Equals("total", StringComparison.OrdinalIgnoreCase))
                 {
                     thresholdStat = ThresholdStatistic.Total;
                 }
@@ -164,23 +154,23 @@ namespace Coverlet.MSbuild.Tasks
 
                 Console.WriteLine(coverageTable.ToStringAlternative());
 
-                thresholdTypeFlags = result.GetThresholdTypesBelowThreshold(summary, _threshold, thresholdTypeFlags, thresholdStat);
+                thresholdTypeFlags = result.GetThresholdTypesBelowThreshold(summary, Threshold, thresholdTypeFlags, thresholdStat);
                 if (thresholdTypeFlags != ThresholdTypeFlags.None)
                 {
                     var exceptionMessageBuilder = new StringBuilder();
                     if ((thresholdTypeFlags & ThresholdTypeFlags.Line) != ThresholdTypeFlags.None)
                     {
-                        exceptionMessageBuilder.AppendLine($"The {thresholdStat.ToString().ToLower()} line coverage is below the specified {_threshold}");
+                        exceptionMessageBuilder.AppendLine($"The {thresholdStat.ToString().ToLower()} line coverage is below the specified {Threshold}");
                     }
 
                     if ((thresholdTypeFlags & ThresholdTypeFlags.Branch) != ThresholdTypeFlags.None)
                     {
-                        exceptionMessageBuilder.AppendLine($"The {thresholdStat.ToString().ToLower()} branch coverage is below the specified {_threshold}");
+                        exceptionMessageBuilder.AppendLine($"The {thresholdStat.ToString().ToLower()} branch coverage is below the specified {Threshold}");
                     }
 
                     if ((thresholdTypeFlags & ThresholdTypeFlags.Method) != ThresholdTypeFlags.None)
                     {
-                        exceptionMessageBuilder.AppendLine($"The {thresholdStat.ToString().ToLower()} method coverage is below the specified {_threshold}");
+                        exceptionMessageBuilder.AppendLine($"The {thresholdStat.ToString().ToLower()} method coverage is below the specified {Threshold}");
                     }
 
                     throw new Exception(exceptionMessageBuilder.ToString());
